refactor: move cryptoOptions to constants file + use Title component
