+++ web-marketplace/src/components/molecules/CreditsAmount/CreditsAmount.mock.tsx
@@ -6,20 +6,3 @@ export const creditVintages = [
   { date: 'Jan 1, 2020 - December 31, 2020', credits: '200' },
   { date: 'Jan 1, 2021 - December 31, 2021', credits: '300' },
 ];
-
-export const cryptoOptions = [
-  {
-    label: 'Retire credits now',
-    description:
-      'These credits will be retired upon purchase and will not be tradeable. Retirement is permanent and non-reversible.',
-    linkTo:
-      'https://guides.regen.network/guides/regen-marketplace-buyers-guides/ecocredits/retire-ecocredits/retirement-certification#individual-entity-credit-retirement.',
-  },
-  {
-    label: 'Buy tradable ecocredits',
-    description:
-      'These credits will be a tradeable asset. They can be retired later via Regen Marketplace.',
-    linkTo:
-      'https://guides.regen.network/guides/regen-marketplace-buyers-guides/ecocredits/retire-ecocredits/retirement-certification#individual-entity-credit-retirement.',
-  },
-];
